refactor: gameplay session 책임을 훅 단위로 분리
